+++ Team8745/src/main/AndroidManifest.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="utf-8"?>
+
+<!-- Note: the actual manifest file used in your APK merges this file with contributions
+     from the modules on which your app depends (such as FtcRobotController, etc).
+     So it won't ultimately be as empty as it might here appear to be :-) -->
+
+<!-- The package name here determines the package for your R class and your BuildConfig class -->
+<manifest
+    package="org.firstinspires.ftc.team8745"
+    xmlns:android="http://schemas.android.com/apk/res/android">
+    <application/>
+</manifest>
