+++ Team8745/src/main/res/values/strings.xml
