+++ Team9374/src/main/AndroidManifest.xml
@@ -6,7 +6,7 @@
 
 <!-- The package name here determines the package for your R class and your BuildConfig class -->
 <manifest
-    package="org.firstinspires.ftc.teamcode"
+    package="org.firstinspires.ftc.team9374"
     xmlns:android="http://schemas.android.com/apk/res/android">
     <application/>
 </manifest>
