+++ settings.gradle
@@ -1,2 +1,3 @@
 include ':FtcRobotController'
-include ':TeamCode'
+include ':Team8745'
+include ':Team9374'
